Module Bil.Apply

Maps BIL operators to bitvectors.

since

1.3

val binop : binop -> word -> word -> word

binop op x y applies the binary operation op to x and y. precondition: the expression BinOp(op,Int x,Int y) shall be well-typed.

val unop : unop -> word -> word

unop op x applies the unary operation op to x. precondition: the expression Unop(op,Int x) shall be well-typed.

val cast : cast -> int -> word -> word

cast t s x casts x using the cast type t to the given size s. precondition: the expression Cast(t,s,Int x) shall be well-typed.