Module Bil.Attribute

BIL attributes.

BIL attributes serve the role of constructor for encoding values as special statements. The attribute defines methods for encoding and decoding values as a string as well as a unique attribute name.

type 'a t

the type of attributes

val declare : ?package:string -> encode:('a -> string) -> decode:(string -> 'a) -> string -> 'a t

declare ?package name ~encode ~decode declares a new attribute.

The attribute package, name pair should be unique. If an attribute with the given name is already registered the registration will fail.