Class Storage.linear

linear storage literally implements operational semantics, but has O(N) lookup and uses space very ineffectively, as it is implemented as a list of assignments.

method load : addr -> word option

load a loads a byte from a a given address a

method save : addr -> word -> 's

save a w stores byte w at address a