Class Storage.sparse

sparse storage is slightly more efficient storage, in comparison with linear. It uses balanced tree data structure, and provides logarithmic lookup and update method.

method load : addr -> word option

load a loads a byte from a a given address a

method save : addr -> word -> 's

save a w stores byte w at address a