Module Bil.Theory

Core Theory specification of BIL.

val parser : (exp, _, stmt) Bap_core_theory.Theory.Parser.t

parser the parser enables reflection of the bil statements into core theory terms. To reflect a bil program prog into the theory Theory, use

let module Parser = Theory.Parser.Make(Theory) in
Parser.run Bil.Theory.parser bil