Class type Bil.storage

An interface to a memory storage.

A storage is a mapping from addresses to bytes. For consistency and efficiency bytes are still reprented with bitvectors.

Storages should not take care of aliasing or endianness, as they are byte addressable. All memory operations are normalized by Bili.

deprecated
Use the Primus Framework.

method load : addr -> word option

load a loads a byte from a a given address a

method save : addr -> word -> 's

save a w stores byte w at address a