Module Std.Biri

BIR Interpreter

deprecated

Use the Primus Framework.

class context : ?⁠main:sub term -> program term -> object ... end

Biri evaluates terms in the context of a whole program (since terms may contain calls and jumps). Biri also tracks for current position inside block, the block and preceding block.

module type S = sig ... end

module Make : functor (M : Monads.Std.Monad.State.S2) -> S with type ('a, 'e) state = ('a, 'e) M.t

include S with type ('a, 'e) S.state = ('a, 'e) Monads.Std.Monad.State.t

type ('a, 'e) state = ('a, 'e) Monads.Std.Monad.State.t

type 'a u = (unit, 'a) state

type 'a r = (Bil.result, 'a) state

module Expi : sig ... end

class 'a t : object ... end

base class for BIR interpreters