Module Std.Context

Base class for evaluation contexts.

All interpreters evaluate terms under a given context, wrapped into a state monad. All context types must be structural subtypes of the Context.t.

The base context is just a mapping from variables to values.

Other than a type Context.t this module has n class t that provides a logarithmic implementation for lookup and update methods.

Since context, for any interpreter must be a structural subtype of Context.t it is not required that this particular should be used. Any implementation that has matching interface will work.

deprecated

Use the Primus Framework

class t : object ... end