Class Exp.state

All visitors provide some information about the current position of the visitor

val exps_stack : exp list

a stack of expr, that are parents for the currently visiting expression

val under_condition : bool

is true if currently visiting entry is executed conditionally