Module Ir.Edge

since in IR the order of edges defines semantics, we provide extra functions

include Graphlib.Std.Edge with type graph = t and type node = node and type t = edge

type t = edge

type node = node

type graph = t

type label

val create : node -> node -> label -> t

create x y l creates an edge connecting nodes x and y labeled with a a given label l

val label : t -> label

label e returns a label of an edge e

val src : t -> node

src e returns a source of an edge e

val dst : t -> node

dst e returns a destination of an edge e

val mem : t -> graph -> bool

mem e g is true if e ∈ E.

val insert : t -> graph -> graph

insert e g returns a graph g' with a set of edges extended with edge e. If src e or dst e wasn't in the set of nodes N, then it is extended as well, so that axioms of graph are preserved.

Postconditions:

          - E(g') = E(g) ∪ {e}.

val update : t -> label -> graph -> graph

update e l g if edge e exists in graph g then return a new graph g' in which edge e is associated with label l. Otherwise return g unchanged.

Postcondition:

          - E(g) ⊆ E(g')
          - N(g) ⊆ N(g')
          - e ∉ E(g) → e ∉ E(g').
          - e ∈ E(g) → ε(g')e = l.

val remove : t -> graph -> graph

remove e g returns a graph g' that doesn't contain edge e.

Postconditions:

          - E(g') = E(g) \ {e}.

include Regular.Std.Opaque.S with type t := t

include Core_kernel.Comparable.S with type t := t

val (>=) : t -> t -> bool

val (<=) : t -> t -> bool

val (=) : t -> t -> bool

val (>) : t -> t -> bool

val (<) : t -> t -> bool

val (<>) : t -> t -> bool

val equal : t -> t -> bool

val min : t -> t -> t

val max : t -> t -> t

val ascending : t -> t -> int

val descending : t -> t -> int

val between : t -> low:t -> high:t -> bool

val clamp_exn : t -> min:t -> max:t -> t

val clamp : t -> min:t -> max:t -> t Base__.Or_error.t

type comparator_witness

val comparator : (t, comparator_witness) Base__.Comparator.comparator

val validate_lbound : min:t Base__.Maybe_bound.t -> t Base__.Validate.check

val validate_ubound : max:t Base__.Maybe_bound.t -> t Base__.Validate.check

val validate_bound : min:t Base__.Maybe_bound.t -> max:t Base__.Maybe_bound.t -> t Base__.Validate.check

module Replace_polymorphic_compare : sig ... end

module Map : sig ... end

module Set : sig ... end

include Core_kernel.Hashable.S with type t := t

val compare : t -> t -> Core_kernel__.Import.int

val hash_fold_t : Ppx_hash_lib.Std.Hash.state -> t -> Ppx_hash_lib.Std.Hash.state

val hash : t -> Ppx_hash_lib.Std.Hash.hash_value

val hashable : t Core_kernel__.Hashtbl.Hashable.t

module Table : sig ... end

module Hash_set : sig ... end

module Hash_queue : sig ... end

val jmps : [ `after | `before ] -> t -> graph -> jmp term seq

jmps dir e g enumerates all jumps (including calls, interrupts, indirects, etc), that occurs before if dir = `before or after if dir = `after an edge e

val edges : [ `after | `before ] -> t -> graph -> t seq

edges dir e g enumerates all edges occurring before of after an edge e in graph g

val jmp : t -> jmp term

jmp e returns a jmp term associated with edge e

val tid : t -> tid

tid e returns a tid of a jmp term that is associated with an edge e

val cond : t -> graph -> exp

cond e g computes a condition expression that is asserted to be true if this branch is taken.

Note: this is not the same as a condition associated with the jmp term itself, it takes into account all conditions preceding the edge.

include Regular.Std.Printable.S with type t := t

val to_string : t -> string

to_string x returns a human-readable representation of x

val str : unit -> t -> string

str () t is formatted output function that matches "%a" conversion format specifier in functions, that prints to string, e.g., sprintf, failwithf, errorf and, surprisingly all Lwt printing function, including Lwt_io.printf and logging (or any other function with type ('a,unit,string,...) formatN`. Example:

Or_error.errorf "type %a is not valid for %a"
  Type.str ty Exp.str exp

val pps : unit -> t -> string

synonym for str

val ppo : Core_kernel.Out_channel.t -> t -> unit

will print to a standard output_channel, useful for using in printf, fprintf, etc.

val pp_seq : Stdlib.Format.formatter -> t Core_kernel.Sequence.t -> unit

prints a sequence of values of type t

this will include pp function from Core that has type t printer, and can be used in Format.printf family of functions

include Core_kernel.Pretty_printer.S with type t := t

val pp : Base__.Formatter.t -> t -> unit