Module Image.KB

Interfaces for working with the Knowledge Base.

module type Loader = sig ... end

Same as Loader, but parameterized with the Knowledge monad.

val register_loader : name:string -> (module Loader) -> unit

register_loader ~name backend registers a new loader that is parameterized with the Knowledge monad.