Module type KB.Loader

Same as Loader, but parameterized with the Knowledge monad.

val from_file : string -> Ogre.doc option Bap_knowledge.knowledge

from_file name loads a file with the given name.

val from_data : Core_kernel.Bigstring.t -> Ogre.doc option Bap_knowledge.knowledge

from_data data loads image from the specified array of bytes.