Module Image.Spec

Interface to the image specification.

val from_arch : arch -> Ogre.doc

from_arch x constructs a minimal specification for the given architecture x.

val slot : (Bap_core_theory.Theory.Unit.cls, Ogre.doc) Bap_core_theory.KB.slot

the slot to access the specification of the uni