Module Std.Integer

Abstract integral type.

This module describes an interface of an integral arithmetic type, as well as the Make functor, that derives a module that implements this interface from a module that provides the minimal (Base) interface

module type Base = sig ... end

The minimal interface of an integer.

module type S = sig ... end

The integer signature.

module Make : functor (T : Base) -> S with type t = T.t

Derive S from the minimal implementation.