Module Std.Interval_tree

Balanced Interval Tree.

Interval trees are used to build efficient mappings from intervals to arbitrary data.

The interval tree may contain overlapping intervals and allows inserting and removing elements.

The interval tree is implemented using AVL trees.

since

1.4

module type Interval = sig ... end

module type S = sig ... end

The Interval Tree Interface.

module Make : functor (Interval : Interval) -> S with type key := Interval.t and type point := Interval.point

Make(Interval) create an abstract interval tree data type that uses abstract Interval.

module type Interval_binable = sig ... end

Binable Abstract Interval.

module type S_binable = sig ... end

Binable Interval Tree.

module Make_binable : functor (Interval : Interval_binable) -> S_binable with type key := Interval.t and type point := Interval.point

Make_binable(Interval) create an abstract interval tree data type that uses abstract Interval and can be serialized via the Binable interface.