Module Interval_tree.Make_binable

Make_binable(Interval) create an abstract interval tree data type that uses abstract Interval and can be serialized via the Binable interface.

Parameters

Interval : Interval_binable

Signature

type 'a t

include Bin_prot.Binable.S1 with type 'a t := 'a t

type 'a t

val bin_shape_t : Bin_prot.Shape.t -> Bin_prot.Shape.t

val bin_size_t : ('a, 'a t) Bin_prot.Size.sizer1

val bin_write_t : ('a, 'a t) Bin_prot.Write.writer1

val bin_read_t : ('a, 'a t) Bin_prot.Read.reader1

val __bin_read_t__ : ('a, int -> 'a t) Bin_prot.Read.reader1

val bin_writer_t : ('a, 'a t) Bin_prot.Type_class.S1.writer

val bin_reader_t : ('a, 'a t) Bin_prot.Type_class.S1.reader

val bin_t : ('a, 'a t) Bin_prot.Type_class.S1.t

val compare : ('a -> 'a -> int) -> 'a t -> 'a t -> int

include Ppx_sexp_conv_lib.Sexpable.S1 with type 'a t := 'a t

type 'a t

val t_of_sexp : (Sexplib0.Sexp.t -> 'a) -> Sexplib0.Sexp.t -> 'a t

val sexp_of_t : ('a -> Sexplib0.Sexp.t) -> 'a t -> Sexplib0.Sexp.t

include S with type 'a t := 'a t

type 'a t

interval tree abstract representation

val sexp_of_t : ('a -> Ppx_sexp_conv_lib.Sexp.t) -> 'a t -> Ppx_sexp_conv_lib.Sexp.t

type key

the interval

type point

an element of the interval

val empty : 'a t

empty x an empty interval tree

val singleton : key -> 'a -> 'a t

singleton k x creates an interval tree that has only one mapping - from the key k to data x

val least : 'a t -> point option

least t returns the least bound of the tree t.

Returns None if t is empty.

val greatest : 'a t -> point option

greatest t returns the greatest bound of the tree t.

Returns None if t is empty.

val min_binding : 'a t -> (key * 'a) option

min_bining t returns the least binding in the tree

val max_binding : 'a t -> (key * 'a) option

max_binding t returns the greatest binding in the tree

val add : 'a t -> key -> 'a -> 'a t

add t k x adds a new binding (k,x) to the mapping.

val dominators : 'a t -> key -> (key * 'a) Core_kernel.Sequence.t

dominators t k returns all intervals and their associated values that include k.

val intersections : 'a t -> key -> (key * 'a) Core_kernel.Sequence.t

intersections t k returns all intervals and their associated values that intersects with k

val intersects : 'a t -> key -> bool

intersects t k is true iff t contains an interval that intersects with k

val dominates : 'a t -> key -> bool

dominates t k is true iff all intervals in t are included in k.

val contains : 'a t -> point -> bool

contains t p is true if p belongs to at least one interval in t

val lookup : 'a t -> point -> (key * 'a) Core_kernel.Sequence.t

lookup t p returns bindings of all intervals that contain the given point

val map : 'a t -> f:('a -> 'b) -> 'b t

map k ~f maps all data values with the function f

val mapi : 'a t -> f:(key -> 'a -> 'b) -> 'b t

mapi k ~f maps all bindings with the function f

val filter : 'a t -> f:('a -> bool) -> 'a t

filter t ~f returns a tree where all elements for which f returned false are removed.

val filter_map : 'a t -> f:('a -> 'b option) -> 'b t

filter t ~f returns a tree where all elements for which f returned None are removed and all others are mapped.

val filter_mapi : 'a t -> f:(key -> 'a -> 'b option) -> 'b t

filter t ~f returns a tree where all elements for which f returned None are removed and all others are mapped.

val remove : 'a t -> key -> 'a t

remove t k removes all bindings to the key k

val remove_intersections : 'a t -> key -> 'a t

remove_intersections t k removes all bindings that intersect with the key k.

val remove_dominators : 'a t -> key -> 'a t

remove_dominators t k removes all bindings that are included (dominated by) in the interval k

val to_sequence : 'a t -> (key * 'a) Core_kernel.Sequence.t

to_sequence t returns all bindings in t

Interval Trees implement common container interface

include Core_kernel.Container.S1 with type 'a t := 'a t

type 'a t

val mem : 'a t -> 'a -> equal:('a -> 'a -> bool) -> bool

Checks whether the provided element is there, using equal.

val length : 'a t -> int

val is_empty : 'a t -> bool

val iter : 'a t -> f:('a -> unit) -> unit

val fold : 'a t -> init:'accum -> f:('accum -> 'a -> 'accum) -> 'accum

fold t ~init ~f returns f (... f (f (f init e1) e2) e3 ...) en, where e1..en are the elements of t

val fold_result : 'a t -> init:'accum -> f:('accum -> 'a -> ('accum, 'e) Base.Result.t) -> ('accum, 'e) Base.Result.t

fold_result t ~init ~f is a short-circuiting version of fold that runs in the Result monad. If f returns an Error _, that value is returned without any additional invocations of f.

val fold_until : 'a t -> init:'accum -> f:('accum -> 'a -> ('accum, 'final) Base__.Container_intf.Continue_or_stop.t) -> finish:('accum -> 'final) -> 'final

fold_until t ~init ~f ~finish is a short-circuiting version of fold. If f returns Stop _ the computation ceases and results in that value. If f returns Continue _, the fold will proceed. If f never returns Stop _, the final result is computed by finish.

Example:

type maybe_negative =
  | Found_negative of int
  | All_nonnegative of { sum : int }

(** [first_neg_or_sum list] returns the first negative number in [list], if any,
    otherwise returns the sum of the list. *)
let first_neg_or_sum =
  List.fold_until ~init:0
    ~f:(fun sum x ->
      if x < 0
      then Stop (Found_negative x)
      else Continue (sum + x))
    ~finish:(fun sum -> All_nonnegative { sum })
;;

let x = first_neg_or_sum [1; 2; 3; 4; 5]
val x : maybe_negative = All_nonnegative {sum = 15}

let y = first_neg_or_sum [1; 2; -3; 4; 5]
val y : maybe_negative = Found_negative -3

val exists : 'a t -> f:('a -> bool) -> bool

Returns true if and only if there exists an element for which the provided function evaluates to true. This is a short-circuiting operation.

val for_all : 'a t -> f:('a -> bool) -> bool

Returns true if and only if the provided function evaluates to true for all elements. This is a short-circuiting operation.

val count : 'a t -> f:('a -> bool) -> int

Returns the number of elements for which the provided function evaluates to true.

val sum : (module Base__.Container_intf.Summable with type t = 'sum) -> 'a t -> f:('a -> 'sum) -> 'sum

Returns the sum of f i for all i in the container.

val find : 'a t -> f:('a -> bool) -> 'a option

Returns as an option the first element for which f evaluates to true.

val find_map : 'a t -> f:('a -> 'b option) -> 'b option

Returns the first evaluation of f that returns Some, and returns None if there is no such element.

val to_list : 'a t -> 'a list

val to_array : 'a t -> 'a array

val min_elt : 'a t -> compare:('a -> 'a -> int) -> 'a option

Returns a minimum (resp maximum) element from the collection using the provided compare function, or None if the collection is empty. In case of a tie, the first element encountered while traversing the collection is returned. The implementation uses fold so it has the same complexity as fold.

val max_elt : 'a t -> compare:('a -> 'a -> int) -> 'a option