Module type Interval_tree.S

The Interval Tree Interface.

Interval tree is a mapping from intervals to arbitrary values. The intervals are allowed to intersect. Thus a single point may belong to more than one interval. Unlike a regular map, when an association is extract by using a key value, the interval tree uses notions of domination and intersection to extract values associated with all intervals that either dominate (i.e., are super sets) or intersects with the provided key. In that sense an interval tree is a multimap.

type 'a t

interval tree abstract representation

val sexp_of_t : ('a -> Ppx_sexp_conv_lib.Sexp.t) -> 'a t -> Ppx_sexp_conv_lib.Sexp.t

type key

the interval

type point

an element of the interval

val empty : 'a t

empty x an empty interval tree

val singleton : key -> 'a -> 'a t

singleton k x creates an interval tree that has only one mapping - from the key k to data x

val least : 'a t -> point option

least t returns the least bound of the tree t.

Returns None if t is empty.

val greatest : 'a t -> point option

greatest t returns the greatest bound of the tree t.

Returns None if t is empty.

val min_binding : 'a t -> (key * 'a) option

min_bining t returns the least binding in the tree

val max_binding : 'a t -> (key * 'a) option

max_binding t returns the greatest binding in the tree

val add : 'a t -> key -> 'a -> 'a t

add t k x adds a new binding (k,x) to the mapping.

val dominators : 'a t -> key -> (key * 'a) Core_kernel.Sequence.t

dominators t k returns all intervals and their associated values that include k.

val intersections : 'a t -> key -> (key * 'a) Core_kernel.Sequence.t

intersections t k returns all intervals and their associated values that intersects with k

val intersects : 'a t -> key -> bool

intersects t k is true iff t contains an interval that intersects with k

val dominates : 'a t -> key -> bool

dominates t k is true iff all intervals in t are included in k.

val contains : 'a t -> point -> bool

contains t p is true if p belongs to at least one interval in t

val lookup : 'a t -> point -> (key * 'a) Core_kernel.Sequence.t

lookup t p returns bindings of all intervals that contain the given point

val map : 'a t -> f:('a -> 'b) -> 'b t

map k ~f maps all data values with the function f

val mapi : 'a t -> f:(key -> 'a -> 'b) -> 'b t

mapi k ~f maps all bindings with the function f

val filter : 'a t -> f:('a -> bool) -> 'a t

filter t ~f returns a tree where all elements for which f returned false are removed.

val filter_map : 'a t -> f:('a -> 'b option) -> 'b t

filter t ~f returns a tree where all elements for which f returned None are removed and all others are mapped.

val filter_mapi : 'a t -> f:(key -> 'a -> 'b option) -> 'b t

filter t ~f returns a tree where all elements for which f returned None are removed and all others are mapped.

val remove : 'a t -> key -> 'a t

remove t k removes all bindings to the key k

val remove_intersections : 'a t -> key -> 'a t

remove_intersections t k removes all bindings that intersect with the key k.

val remove_dominators : 'a t -> key -> 'a t

remove_dominators t k removes all bindings that are included (dominated by) in the interval k

val to_sequence : 'a t -> (key * 'a) Core_kernel.Sequence.t

to_sequence t returns all bindings in t

Interval Trees implement common container interface

include Core_kernel.Container.S1 with type 'a t := 'a t

val mem : 'a t -> 'a -> equal:('a -> 'a -> bool) -> bool

val length : 'a t -> int

val is_empty : 'a t -> bool

val iter : 'a t -> f:('a -> unit) -> unit

val fold : 'a t -> init:'accum -> f:('accum -> 'a -> 'accum) -> 'accum

val fold_result : 'a t -> init:'accum -> f:('accum -> 'a -> ('accum, 'e) Base__.Result.t) -> ('accum, 'e) Base__.Result.t

val fold_until : 'a t -> init:'accum -> f:('accum -> 'a -> ('accum, 'final) Base__.Container_intf.Continue_or_stop.t) -> finish:('accum -> 'final) -> 'final

val exists : 'a t -> f:('a -> bool) -> bool

val for_all : 'a t -> f:('a -> bool) -> bool

val count : 'a t -> f:('a -> bool) -> int

val sum : (module Base__.Container_intf.Summable with type t = 'sum) -> 'a t -> f:('a -> 'sum) -> 'sum

val find : 'a t -> f:('a -> bool) -> 'a option

val find_map : 'a t -> f:('a -> 'b option) -> 'b option

val to_list : 'a t -> 'a list

val to_array : 'a t -> 'a array

val min_elt : 'a t -> compare:('a -> 'a -> int) -> 'a option

val max_elt : 'a t -> compare:('a -> 'a -> int) -> 'a option