Module Std.Kind

Kinds of instructions

type branch = [

| `Conditional_branch

| `Unconditional_branch

| `Indirect_branch

]

val bin_shape_branch : Bin_prot.Shape.t

val bin_size_branch : branch Bin_prot.Size.sizer

val bin_write_branch : branch Bin_prot.Write.writer

val bin_writer_branch : branch Bin_prot.Type_class.writer

val bin_read_branch : branch Bin_prot.Read.reader

val __bin_read_branch__ : (int -> branch) Bin_prot.Read.reader

val bin_reader_branch : branch Bin_prot.Type_class.reader

val bin_branch : branch Bin_prot.Type_class.t

val compare_branch : branch -> branch -> int

val all_of_branch : branch list

val sexp_of_branch : branch -> Ppx_sexp_conv_lib.Sexp.t

val branch_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> branch

val __branch_of_sexp__ : Ppx_sexp_conv_lib.Sexp.t -> branch

type affecting_control = [

| branch

| `Return

| `Call

| `Barrier

| `Terminator

| `May_affect_control_flow

]

val bin_shape_affecting_control : Bin_prot.Shape.t

val bin_size_affecting_control : affecting_control Bin_prot.Size.sizer

val bin_write_affecting_control : affecting_control Bin_prot.Write.writer

val bin_writer_affecting_control : affecting_control Bin_prot.Type_class.writer

val bin_read_affecting_control : affecting_control Bin_prot.Read.reader

val __bin_read_affecting_control__ : (int -> affecting_control) Bin_prot.Read.reader

val bin_reader_affecting_control : affecting_control Bin_prot.Type_class.reader

val bin_affecting_control : affecting_control Bin_prot.Type_class.t

val compare_affecting_control : affecting_control -> affecting_control -> int

val all_of_affecting_control : affecting_control list

val sexp_of_affecting_control : affecting_control -> Ppx_sexp_conv_lib.Sexp.t

val affecting_control_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> affecting_control

val __affecting_control_of_sexp__ : Ppx_sexp_conv_lib.Sexp.t -> affecting_control

type having_side_effect = [

| `May_load

| `May_store

]

val bin_shape_having_side_effect : Bin_prot.Shape.t

val bin_size_having_side_effect : having_side_effect Bin_prot.Size.sizer

val bin_write_having_side_effect : having_side_effect Bin_prot.Write.writer

val bin_writer_having_side_effect : having_side_effect Bin_prot.Type_class.writer

val bin_read_having_side_effect : having_side_effect Bin_prot.Read.reader

val __bin_read_having_side_effect__ : (int -> having_side_effect) Bin_prot.Read.reader

val bin_reader_having_side_effect : having_side_effect Bin_prot.Type_class.reader

val bin_having_side_effect : having_side_effect Bin_prot.Type_class.t

val compare_having_side_effect : having_side_effect -> having_side_effect -> int

val all_of_having_side_effect : having_side_effect list

val sexp_of_having_side_effect : having_side_effect -> Ppx_sexp_conv_lib.Sexp.t

val having_side_effect_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> having_side_effect

val __having_side_effect_of_sexp__ : Ppx_sexp_conv_lib.Sexp.t -> having_side_effect

type t = [

| affecting_control

| having_side_effect

]

include Bin_prot.Binable.S with type t := t

type t

include Bin_prot.Binable.S_only_functions with type t := t

type t

val bin_size_t : t Bin_prot.Size.sizer

val bin_write_t : t Bin_prot.Write.writer

val bin_read_t : t Bin_prot.Read.reader

val __bin_read_t__ : (int -> t) Bin_prot.Read.reader

This function only needs implementation if t exposed to be a polymorphic variant. Despite what the type reads, this does *not* produce a function after reading; instead it takes the constructor tag (int) before reading and reads the rest of the variant t afterwards.

val bin_shape_t : Bin_prot.Shape.t

val bin_writer_t : t Bin_prot.Type_class.writer

val bin_reader_t : t Bin_prot.Type_class.reader

val bin_t : t Bin_prot.Type_class.t

val compare : t -> t -> int

val all : t list

val sexp_of_t : t -> Ppx_sexp_conv_lib.Sexp.t

val t_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> t

val __t_of_sexp__ : Ppx_sexp_conv_lib.Sexp.t -> t