Module Project.State

The core state of the project.

type t = state

the abstract type for the project state. See Project.state.

val disassembly : t -> Disasm.Driver.state

disassembly state contains all disassembled instructions, as well as their connection. To build control-flow graphs or to explore the graph structure, use Disasm.Driver.explore.

val subroutines : t -> Disasm.Subroutines.t

subroutines state returns the partition of the set of disassembled instructions into a set of subroutines.

val slot : ( Bap_core_theory.Theory.Unit.cls, state ) Bap_core_theory.KB.slot

the slot of a unit object that stores the state of disassembly