Class Stmt.state

All visitors provide some information about the current position of the visitor

val preds : stmt list

the stack of stmts that was already visited, with the last on the top. Not including the currently visiting stmt.

val succs : stmt list

stmts that are not yet visited

val stmts_stack : stmt list

a stack of stmts that are parents for the currently visiting entity. The top one is the one that we're currently visiting.

val in_jmp : bool

is true if we're visiting expression that is a jump target

val in_move : bool

is true if we're visiting expression that is on the left or right side of the assignment.

val in_loop : bool

is true if currently visiting expression or statement is executed under loop.