Module Std.Sub

Subroutine.

type t = sub term

Subroutine is a set of blocks. The first block of a function is considered an entry block.

val create : ?⁠args:arg term list -> ?⁠blks:blk term list -> ?⁠tid:tid -> ?⁠name:string -> unit -> t

create ?name () creates a new subroutine.

Creates a subroutine that includes given arguments and blocks. The order of the terms is preserved with the first block being the entry block. No references between blocks are added, so the blocks shall be correctly linked and be reachable from the entry block.

If tid is not specied then a fresh one is generated. if name is not specified then a fresh name is derived from the tid.

since

2.3.0 has the [args] optional parameter

since

2.3.0 has the [blks] optional parameter

val lift : block -> cfg -> sub term

lift entry takes an basic block of assembler instructions, as an entry and lifts it to the subroutine term.

val name : t -> string

name sub returns a subroutine name

val with_name : t -> string -> t

updates subroutine name

val ssa : t -> t

ssa sub returns sub in SSA form. If program is already in SSA, then do nothing (see also is_ssa). The underlying algorithm produces a semi-pruned SSA form. To represent different versions of the same variable we use variable versions. Any definition of a variable increases its version number. So, the zero version is reserved for variables that weren't defined before the first use.

val is_ssa : t -> bool

is_ssa sub is true if sub was transformed into an SSA form. This is O(1) predicate that doesn't really check, that a subroutine is in an SSA form, so it is a responsibility of a user to preserve the SSA form on any transformation.

val free_vars : t -> Var.Set.t

free_vars sub computes a set of variables that are free in a given subroutine sub. The variable is considered free if it is used before defined or is not locally bound. If sub is in an SSA form, then the set is computed trivially, thanks to a naming scheme. If program is not in an SSA form, then a BFS on a dominators tree is used.

val to_graph : t -> Graphs.Tid.t

to_graph sub builds a graph of subroutine sub.

Graph nodes are block term identifiers and edges are labeled with term identifiers of the jmp terms that correspond to the given edge.

since

2.1 the returned graph contains two pseudo-nodes

Graphs.Tid.start and Graphs.Tid.exit so that all nodes that has in-degree 0 or that start a strongly connected component are connected to the start node (the same for exit but on the reversed graph.

Edges from start to other nodes are labeled with the Graphs.Tid.start tid.

Edges from nodes to the exit node are labeled with the Graphs.Tid.exit tid.

val to_cfg : t -> Graphs.Ir.t

to_cfg sub builds a graph representation of a subroutine sub. All graph operations are mapped to corresponding Term operations. See Graphlib.Ir for more information.

val of_cfg : Graphs.Ir.t -> t

of_cfg cfg extracts a sub term from a given graph cfg. Since Graphlib.Ir module builds term incrementally this operation is just a projection, i.e., it has O(0) complexity.

val compute_liveness : t -> (tid, Var.Set.t) Graphlib.Std.Solution.t

compute_liveness sub computes a set of live variables for each block.

For a block b and solution s = compute_liveness sub, Solution.get s (Term.tid b) is a set of variables that are live at the _exit_ from this block.

A set of variables that are live (free) in the whole subroutine is the set of variables that are live at the Graphs.Tid.start node.

since

2.1

val aliases : string list tag

other names for the given subroutine.

val const : unit tag

A subroutine doesn't examine any values except its arguments, and have no effects except the return value. Basically this is just slightly more strict class than the pure attribute below, since function is not allowed to read global memory. Note that a function that has pointer arguments and examines the data pointed to is not const. Likewise, a function that calls a non-const function usually is not be const. It does not make sense for a const function to return void

val pure : unit tag

A subroutine have no effects except the return value and their return value depends only on the parameters and/or global variables.

val stub : unit tag

A subroutine is a stub

val extern : unit tag

A subroutine is visible outside of the compilation unit

val leaf : unit tag

a subroutine doesn't contain any calls in any disguise, i.e., no longjmps, indirect calls, exceptions, etc.

val malloc : unit tag

A subroutine is malloc-like, i.e., the pointer P returned by the subroutine cannot alias any other pointer valid when the function returns, and moreover no pointers to valid objects occur in any storage addressed by P.

val noreturn : unit tag

A subroutine will not return (either loop infinitely or abort a program)

val returns_twice : unit tag

A subroutine may return more than one time. Examples of such functions are setjmp and vfork

val nothrow : unit tag

A subroutine doesn't throw exceptions

val entry_point : unit tag

a subroutine is the binary entry point

module Builder : sig ... end

Subroutine builder

val pp_slots : string list -> Stdlib.Format.formatter -> t -> unit

pp_slots names prints slots that are in names.

include Regular.Std.Regular.S with type t := t

type t

include Core_kernel.Bin_prot.Binable.S with type Sub.t := t

val bin_size_t : t Bin_prot.Size.sizer

val bin_write_t : t Bin_prot.Write.writer

val bin_read_t : t Bin_prot.Read.reader

val __bin_read_t__ : (int -> t) Bin_prot.Read.reader

val bin_shape_t : Bin_prot.Shape.t

val bin_writer_t : t Bin_prot.Type_class.writer

val bin_reader_t : t Bin_prot.Type_class.reader

val bin_t : t Bin_prot.Type_class.t

include Ppx_sexp_conv_lib.Sexpable.S with type Sub.t := t

val t_of_sexp : Sexplib0__.Sexp.t -> t

val sexp_of_t : t -> Sexplib0__.Sexp.t

val compare : t -> t -> int

include Regular.Std.Printable.S with type t := t

type t

type of printable

val to_string : t -> string

to_string x returns a human-readable representation of x

val str : unit -> t -> string

str () t is formatted output function that matches "%a" conversion format specifier in functions, that prints to string, e.g., sprintf, failwithf, errorf and, surprisingly all Lwt printing function, including Lwt_io.printf and logging (or any other function with type ('a,unit,string,...) formatN`. Example:

Or_error.errorf "type %a is not valid for %a"
  Type.str ty Exp.str exp

val pps : unit -> t -> string

synonym for str

val ppo : Core_kernel.Out_channel.t -> t -> unit

will print to a standard output_channel, useful for using in printf, fprintf, etc.

val pp_seq : Stdlib.Format.formatter -> t Core_kernel.Sequence.t -> unit

prints a sequence of values of type t

this will include pp function from Core that has type t printer, and can be used in Format.printf family of functions

include Core_kernel.Pretty_printer.S with type Sub.t := t

val pp : Base__.Formatter.t -> t -> unit

include Core_kernel.Comparable.S_binable with type Sub.t := t

val (>=) : t -> t -> bool

val (<=) : t -> t -> bool

val (=) : t -> t -> bool

val (>) : t -> t -> bool

val (<) : t -> t -> bool

val (<>) : t -> t -> bool

val equal : t -> t -> bool

val compare : t -> t -> int

val min : t -> t -> t

val max : t -> t -> t

val ascending : t -> t -> int

val descending : t -> t -> int

val between : t -> low:t -> high:t -> bool

val clamp_exn : t -> min:t -> max:t -> t

val clamp : t -> min:t -> max:t -> t Base__.Or_error.t

type comparator_witness

val validate_lbound : min:t Base__.Maybe_bound.t -> t Base__.Validate.check

val validate_ubound : max:t Base__.Maybe_bound.t -> t Base__.Validate.check

val validate_bound : min:t Base__.Maybe_bound.t -> max:t Base__.Maybe_bound.t -> t Base__.Validate.check

module Replace_polymorphic_compare : sig ... end

val comparator : (t, comparator_witness) Core_kernel__.Comparator.comparator

module Map : sig ... end

module Set : sig ... end

include Core_kernel.Hashable.S_binable with type Sub.t := t

val hash_fold_t : Ppx_hash_lib.Std.Hash.state -> t -> Ppx_hash_lib.Std.Hash.state

val hash : t -> Ppx_hash_lib.Std.Hash.hash_value

val hashable : t Core_kernel__.Hashtbl.Hashable.t

module Table : sig ... end

module Hash_set : sig ... end

module Hash_queue : sig ... end

include Regular.Std.Data.S with type t := t

type t

type constructor

type info = string * [ `Ver of string ] * string option

name,Ver v,desc information attached to a particular reader or writer.

val version : string

Data representation version. After any change in data representation the version should be increased.

Serializers that are derived from a data representation must have the same version as a version of the data structure, from which it is derived. This kind of serializers can only read and write data of the same version.

Other serializers can actually read and write data independent on its representation version. A serializer, that can't store data of current version simply shouldn't be added to a set of serializers.

It is assumed, that if a reader and a writer has the same name and version, then whatever was written by the writer should be readable by the reader. The round-trip equality is not required, thus it is acceptable if some information is lost.

It is also possible, that a reader and a writer that has the same name are compatible. In that case it is recommended to use semantic versioning.

val size_in_bytes : ?⁠ver:string -> ?⁠fmt:string -> t -> int

size_in_bytes ?ver ?fmt datum returns the amount of bytes that is needed to represent datum in the given format and version

val of_bytes : ?⁠ver:string -> ?⁠fmt:string -> Regular.Std.bytes -> t

of_bytes ?ver ?fmt bytes deserializes a value from bytes.

val to_bytes : ?⁠ver:string -> ?⁠fmt:string -> t -> Regular.Std.bytes

to_bytes ?ver ?fmt datum serializes a datum to a sequence of bytes.

val blit_to_bytes : ?⁠ver:string -> ?⁠fmt:string -> Regular.Std.bytes -> t -> int -> unit

blit_to_bytes ?ver ?fmt buffer datum offset copies a serialized representation of datum into a buffer, starting from the offset.

val of_bigstring : ?⁠ver:string -> ?⁠fmt:string -> Core_kernel.bigstring -> t

of_bigstring ?ver ?fmt buf deserializes a datum from bigstring

val to_bigstring : ?⁠ver:string -> ?⁠fmt:string -> t -> Core_kernel.bigstring

of_bigstring ?ver ?fmt datum serializes a datum to a sequence of bytes represented as bigstring

val blit_to_bigstring : ?⁠ver:string -> ?⁠fmt:string -> Core_kernel.bigstring -> t -> int -> unit

blit_to_bigstring ?ver ?fmt buffer datum offset copies a serialized representation of datum into a buffer, starting from offset.

module Io : sig ... end

Input/Output functions for the given datum.

module Cache : sig ... end

Data cache.

val add_reader : ?⁠desc:string -> ver:string -> string -> t Regular.Std.reader -> unit

add_reader ?desc ~ver name reader registers a new reader with a provided name, version ver and optional description desc

val add_writer : ?⁠desc:string -> ver:string -> string -> t Regular.Std.writer -> unit

add_writer ?desc ~ver name writer registers a new writer with a provided name, version ver and optional description desc

val available_readers : unit -> info list

available_reader () lists available readers for the data type

val default_reader : unit -> info

default_reader returns information about default reader

val set_default_reader : ?⁠ver:string -> string -> unit

set_default_reader ?ver name sets new default reader. If version is not specified then the latest available version is used. Raises an exception if a reader with a given name doesn't exist.

val with_reader : ?⁠ver:string -> string -> (unit -> 'a) -> 'a

with_reader ?ver name operation temporary sets a default reader to a reader with a specified name and version. The default reader is restored after operation is finished.

val available_writers : unit -> info list

available_writer () lists available writers for the data type

val default_writer : unit -> info

default_writer returns information about the default writer

val set_default_writer : ?⁠ver:string -> string -> unit

set_default_writer ?ver name sets new default writer. If version is not specified then the latest available version is used. Raises an exception if a writer with a given name doesn't exist.

val with_writer : ?⁠ver:string -> string -> (unit -> 'a) -> 'a

with_writer ?ver name operation temporary sets a default writer to a writer with a specified name and version. The default writer is restored after operation is finished.

val default_printer : unit -> info option

default_writer optionally returns an information about default printer

val set_default_printer : ?⁠ver:string -> string -> unit

set_default_printer ?ver name sets new default printer. If version is not specified then the latest available version is used. Raises an exception if a printer with a given name doesn't exist.

val with_printer : ?⁠ver:string -> string -> (unit -> 'a) -> 'a

with_printer ?ver name operation temporary sets a default printer to a printer with a specified name and version. The default printer is restored after operation is finished.

Low level access to serializers

val find_reader : ?⁠ver:string -> string -> t Regular.Std.reader option

find_reader ?ver name lookups a reader with a given name. If version is not specified, then a reader with maximum version is returned.

val find_writer : ?⁠ver:string -> string -> t Regular.Std.writer option

find_writer ?ver name lookups a writer with a given name. If version is not specified, then a writer with maximum version is returned.