Class Taint.context

method taint_reg : Bil.result -> set -> 's

taint result with the given set of taints

method taint_ptr : addr -> size -> set -> 's

taint memory region addr, addr+size with the given set of taints

method reg_taints : Bil.result -> set

returns a set of taints associated with a given result of computation

method ptr_taints : addr -> set

returns a set of taints associated with a given address

method all_taints : set

returns all known taints.