Module type Std.Target

Abstract interface for all targets.

Each target supported by BAP implements this interface. To get access to the implementation use target_of_arch function. Code written using this interface is cross-platform, i.e., target agnostic. If you want to write target-specific code, then use directly corresponding libraries: ARM, IA32, AMD64.

module CPU : CPU

access to the CPU variables.

val lift : lifter

lift mem insn lifts provided instruction to BIL. Usually you do not need to call this function directly, as disassemble function will do the lifting.

postcondition: the returned BIL code is well-typed.