Module Std.Arch

Architecture

type x86 = [

| `x86

| `x86_64

]

val bin_x86 : x86 Bin_prot.Type_class.t

val bin_read_x86 : x86 Bin_prot.Read.reader

val __bin_read_x86__ : (int -> x86) Bin_prot.Read.reader

val bin_reader_x86 : x86 Bin_prot.Type_class.reader

val bin_size_x86 : x86 Bin_prot.Size.sizer

val bin_write_x86 : x86 Bin_prot.Write.writer

val bin_writer_x86 : x86 Bin_prot.Type_class.writer

val bin_shape_x86 : Bin_prot.Shape.t

val compare_x86 : x86 -> x86 -> int

val all_of_x86 : x86 list

val x86_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> x86

val __x86_of_sexp__ : Ppx_sexp_conv_lib.Sexp.t -> x86

val sexp_of_x86 : x86 -> Ppx_sexp_conv_lib.Sexp.t

type arm = [

| `armv4

| `armv5

| `armv6

| `armv7

]

val bin_arm : arm Bin_prot.Type_class.t

val bin_read_arm : arm Bin_prot.Read.reader

val __bin_read_arm__ : (int -> arm) Bin_prot.Read.reader

val bin_reader_arm : arm Bin_prot.Type_class.reader

val bin_size_arm : arm Bin_prot.Size.sizer

val bin_write_arm : arm Bin_prot.Write.writer

val bin_writer_arm : arm Bin_prot.Type_class.writer

val bin_shape_arm : Bin_prot.Shape.t

val compare_arm : arm -> arm -> int

val all_of_arm : arm list

val arm_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> arm

val __arm_of_sexp__ : Ppx_sexp_conv_lib.Sexp.t -> arm

val sexp_of_arm : arm -> Ppx_sexp_conv_lib.Sexp.t

type armeb = [

| `armv4eb

| `armv5eb

| `armv6eb

| `armv7eb

]

val bin_armeb : armeb Bin_prot.Type_class.t

val bin_read_armeb : armeb Bin_prot.Read.reader

val __bin_read_armeb__ : (int -> armeb) Bin_prot.Read.reader

val bin_reader_armeb : armeb Bin_prot.Type_class.reader

val bin_size_armeb : armeb Bin_prot.Size.sizer

val bin_write_armeb : armeb Bin_prot.Write.writer

val bin_writer_armeb : armeb Bin_prot.Type_class.writer

val bin_shape_armeb : Bin_prot.Shape.t

val compare_armeb : armeb -> armeb -> int

val all_of_armeb : armeb list

val armeb_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> armeb

val __armeb_of_sexp__ : Ppx_sexp_conv_lib.Sexp.t -> armeb

val sexp_of_armeb : armeb -> Ppx_sexp_conv_lib.Sexp.t

type thumb = [

| `thumbv4

| `thumbv5

| `thumbv6

| `thumbv7

]

val bin_thumb : thumb Bin_prot.Type_class.t

val bin_read_thumb : thumb Bin_prot.Read.reader

val __bin_read_thumb__ : (int -> thumb) Bin_prot.Read.reader

val bin_reader_thumb : thumb Bin_prot.Type_class.reader

val bin_size_thumb : thumb Bin_prot.Size.sizer

val bin_write_thumb : thumb Bin_prot.Write.writer

val bin_writer_thumb : thumb Bin_prot.Type_class.writer

val bin_shape_thumb : Bin_prot.Shape.t

val compare_thumb : thumb -> thumb -> int

val all_of_thumb : thumb list

val thumb_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> thumb

val __thumb_of_sexp__ : Ppx_sexp_conv_lib.Sexp.t -> thumb

val sexp_of_thumb : thumb -> Ppx_sexp_conv_lib.Sexp.t

type thumbeb = [

| `thumbv4eb

| `thumbv5eb

| `thumbv6eb

| `thumbv7eb

]

val bin_thumbeb : thumbeb Bin_prot.Type_class.t

val bin_read_thumbeb : thumbeb Bin_prot.Read.reader

val __bin_read_thumbeb__ : (int -> thumbeb) Bin_prot.Read.reader

val bin_reader_thumbeb : thumbeb Bin_prot.Type_class.reader

val bin_size_thumbeb : thumbeb Bin_prot.Size.sizer

val bin_write_thumbeb : thumbeb Bin_prot.Write.writer

val bin_writer_thumbeb : thumbeb Bin_prot.Type_class.writer

val bin_shape_thumbeb : Bin_prot.Shape.t

val compare_thumbeb : thumbeb -> thumbeb -> int

val all_of_thumbeb : thumbeb list

val thumbeb_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> thumbeb

val __thumbeb_of_sexp__ : Ppx_sexp_conv_lib.Sexp.t -> thumbeb

val sexp_of_thumbeb : thumbeb -> Ppx_sexp_conv_lib.Sexp.t

type aarch64 = [

| `aarch64

| `aarch64_be

]

val bin_aarch64 : aarch64 Bin_prot.Type_class.t

val bin_read_aarch64 : aarch64 Bin_prot.Read.reader

val __bin_read_aarch64__ : (int -> aarch64) Bin_prot.Read.reader

val bin_reader_aarch64 : aarch64 Bin_prot.Type_class.reader

val bin_size_aarch64 : aarch64 Bin_prot.Size.sizer

val bin_write_aarch64 : aarch64 Bin_prot.Write.writer

val bin_writer_aarch64 : aarch64 Bin_prot.Type_class.writer

val bin_shape_aarch64 : Bin_prot.Shape.t

val compare_aarch64 : aarch64 -> aarch64 -> int

val all_of_aarch64 : aarch64 list

val aarch64_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> aarch64

val __aarch64_of_sexp__ : Ppx_sexp_conv_lib.Sexp.t -> aarch64

val sexp_of_aarch64 : aarch64 -> Ppx_sexp_conv_lib.Sexp.t

type ppc = [

| `ppc

| `ppc64

| `ppc64le

]

val bin_ppc : ppc Bin_prot.Type_class.t

val bin_read_ppc : ppc Bin_prot.Read.reader

val __bin_read_ppc__ : (int -> ppc) Bin_prot.Read.reader

val bin_reader_ppc : ppc Bin_prot.Type_class.reader

val bin_size_ppc : ppc Bin_prot.Size.sizer

val bin_write_ppc : ppc Bin_prot.Write.writer

val bin_writer_ppc : ppc Bin_prot.Type_class.writer

val bin_shape_ppc : Bin_prot.Shape.t

val compare_ppc : ppc -> ppc -> int

val all_of_ppc : ppc list

val ppc_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> ppc

val __ppc_of_sexp__ : Ppx_sexp_conv_lib.Sexp.t -> ppc

val sexp_of_ppc : ppc -> Ppx_sexp_conv_lib.Sexp.t

type mips = [

| `mips

| `mipsel

| `mips64

| `mips64el

]

val bin_mips : mips Bin_prot.Type_class.t

val bin_read_mips : mips Bin_prot.Read.reader

val __bin_read_mips__ : (int -> mips) Bin_prot.Read.reader

val bin_reader_mips : mips Bin_prot.Type_class.reader

val bin_size_mips : mips Bin_prot.Size.sizer

val bin_write_mips : mips Bin_prot.Write.writer

val bin_writer_mips : mips Bin_prot.Type_class.writer

val bin_shape_mips : Bin_prot.Shape.t

val compare_mips : mips -> mips -> int

val all_of_mips : mips list

val mips_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> mips

val __mips_of_sexp__ : Ppx_sexp_conv_lib.Sexp.t -> mips

val sexp_of_mips : mips -> Ppx_sexp_conv_lib.Sexp.t

type sparc = [

| `sparc

| `sparcv9

]

val bin_sparc : sparc Bin_prot.Type_class.t

val bin_read_sparc : sparc Bin_prot.Read.reader

val __bin_read_sparc__ : (int -> sparc) Bin_prot.Read.reader

val bin_reader_sparc : sparc Bin_prot.Type_class.reader

val bin_size_sparc : sparc Bin_prot.Size.sizer

val bin_write_sparc : sparc Bin_prot.Write.writer

val bin_writer_sparc : sparc Bin_prot.Type_class.writer

val bin_shape_sparc : Bin_prot.Shape.t

val compare_sparc : sparc -> sparc -> int

val all_of_sparc : sparc list

val sparc_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> sparc

val __sparc_of_sexp__ : Ppx_sexp_conv_lib.Sexp.t -> sparc

val sexp_of_sparc : sparc -> Ppx_sexp_conv_lib.Sexp.t

type nvptx = [

| `nvptx

| `nvptx64

]

val bin_nvptx : nvptx Bin_prot.Type_class.t

val bin_read_nvptx : nvptx Bin_prot.Read.reader

val __bin_read_nvptx__ : (int -> nvptx) Bin_prot.Read.reader

val bin_reader_nvptx : nvptx Bin_prot.Type_class.reader

val bin_size_nvptx : nvptx Bin_prot.Size.sizer

val bin_write_nvptx : nvptx Bin_prot.Write.writer

val bin_writer_nvptx : nvptx Bin_prot.Type_class.writer

val bin_shape_nvptx : Bin_prot.Shape.t

val compare_nvptx : nvptx -> nvptx -> int

val all_of_nvptx : nvptx list

val nvptx_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> nvptx

val __nvptx_of_sexp__ : Ppx_sexp_conv_lib.Sexp.t -> nvptx

val sexp_of_nvptx : nvptx -> Ppx_sexp_conv_lib.Sexp.t

type hexagon = [

| `hexagon

]

val bin_hexagon : hexagon Bin_prot.Type_class.t

val bin_read_hexagon : hexagon Bin_prot.Read.reader

val __bin_read_hexagon__ : (int -> hexagon) Bin_prot.Read.reader

val bin_reader_hexagon : hexagon Bin_prot.Type_class.reader

val bin_size_hexagon : hexagon Bin_prot.Size.sizer

val bin_write_hexagon : hexagon Bin_prot.Write.writer

val bin_writer_hexagon : hexagon Bin_prot.Type_class.writer

val bin_shape_hexagon : Bin_prot.Shape.t

val compare_hexagon : hexagon -> hexagon -> int

val all_of_hexagon : hexagon list

val hexagon_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> hexagon

val __hexagon_of_sexp__ : Ppx_sexp_conv_lib.Sexp.t -> hexagon

val sexp_of_hexagon : hexagon -> Ppx_sexp_conv_lib.Sexp.t

type r600 = [

| `r600

]

val bin_r600 : r600 Bin_prot.Type_class.t

val bin_read_r600 : r600 Bin_prot.Read.reader

val __bin_read_r600__ : (int -> r600) Bin_prot.Read.reader

val bin_reader_r600 : r600 Bin_prot.Type_class.reader

val bin_size_r600 : r600 Bin_prot.Size.sizer

val bin_write_r600 : r600 Bin_prot.Write.writer

val bin_writer_r600 : r600 Bin_prot.Type_class.writer

val bin_shape_r600 : Bin_prot.Shape.t

val compare_r600 : r600 -> r600 -> int

val all_of_r600 : r600 list

val r600_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> r600

val __r600_of_sexp__ : Ppx_sexp_conv_lib.Sexp.t -> r600

val sexp_of_r600 : r600 -> Ppx_sexp_conv_lib.Sexp.t

type systemz = [

| `systemz

]

val bin_systemz : systemz Bin_prot.Type_class.t

val bin_read_systemz : systemz Bin_prot.Read.reader

val __bin_read_systemz__ : (int -> systemz) Bin_prot.Read.reader

val bin_reader_systemz : systemz Bin_prot.Type_class.reader

val bin_size_systemz : systemz Bin_prot.Size.sizer

val bin_write_systemz : systemz Bin_prot.Write.writer

val bin_writer_systemz : systemz Bin_prot.Type_class.writer

val bin_shape_systemz : Bin_prot.Shape.t

val compare_systemz : systemz -> systemz -> int

val all_of_systemz : systemz list

val systemz_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> systemz

val __systemz_of_sexp__ : Ppx_sexp_conv_lib.Sexp.t -> systemz

val sexp_of_systemz : systemz -> Ppx_sexp_conv_lib.Sexp.t

type xcore = [

| `xcore

]

val bin_xcore : xcore Bin_prot.Type_class.t

val bin_read_xcore : xcore Bin_prot.Read.reader

val __bin_read_xcore__ : (int -> xcore) Bin_prot.Read.reader

val bin_reader_xcore : xcore Bin_prot.Type_class.reader

val bin_size_xcore : xcore Bin_prot.Size.sizer

val bin_write_xcore : xcore Bin_prot.Write.writer

val bin_writer_xcore : xcore Bin_prot.Type_class.writer

val bin_shape_xcore : Bin_prot.Shape.t

val compare_xcore : xcore -> xcore -> int

val all_of_xcore : xcore list

val xcore_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> xcore

val __xcore_of_sexp__ : Ppx_sexp_conv_lib.Sexp.t -> xcore

val sexp_of_xcore : xcore -> Ppx_sexp_conv_lib.Sexp.t

type unknown = [

| `unknown

]

val bin_unknown : unknown Bin_prot.Type_class.t

val bin_read_unknown : unknown Bin_prot.Read.reader

val __bin_read_unknown__ : (int -> unknown) Bin_prot.Read.reader

val bin_reader_unknown : unknown Bin_prot.Type_class.reader

val bin_size_unknown : unknown Bin_prot.Size.sizer

val bin_write_unknown : unknown Bin_prot.Write.writer

val bin_writer_unknown : unknown Bin_prot.Type_class.writer

val bin_shape_unknown : Bin_prot.Shape.t

val compare_unknown : unknown -> unknown -> int

val all_of_unknown : unknown list

val unknown_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> unknown

val __unknown_of_sexp__ : Ppx_sexp_conv_lib.Sexp.t -> unknown

val sexp_of_unknown : unknown -> Ppx_sexp_conv_lib.Sexp.t

type t = [

| aarch64

| arm

| armeb

| thumb

| thumbeb

| hexagon

| mips

| nvptx

| ppc

| r600

| sparc

| systemz

| x86

| xcore

| unknown

]

val bin_t : t Bin_prot.Type_class.t

val bin_read_t : t Bin_prot.Read.reader

val __bin_read_t__ : (int -> t) Bin_prot.Read.reader

val bin_reader_t : t Bin_prot.Type_class.reader

val bin_size_t : t Bin_prot.Size.sizer

val bin_write_t : t Bin_prot.Write.writer

val bin_writer_t : t Bin_prot.Type_class.writer

val bin_shape_t : Bin_prot.Shape.t

val compare : t -> t -> int

val all : t list

val t_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> t

val __t_of_sexp__ : Ppx_sexp_conv_lib.Sexp.t -> t

val sexp_of_t : t -> Ppx_sexp_conv_lib.Sexp.t

val of_string : string -> t option

of_string s will try to be clever and to capture all commonly known synonyms, e.g., of_string "i686" will work

val addr_size : t -> addr_size

addr_size arch returns an address size for a a given arch

val endian : t -> endian

endian arch returns a word endianness of the arch

val slot : (Bap_core_theory.Theory.program, t) Bap_knowledge.Knowledge.slot

the architecture (ISA) of a program.

arch type implements Regular interface

include Regular.Std.Regular.S with type t := t

type t

val bin_t : t Bin_prot.Type_class.t

val bin_read_t : t Bin_prot.Read.reader

val __bin_read_t__ : (int -> t) Bin_prot.Read.reader

val bin_reader_t : t Bin_prot.Type_class.reader

val bin_size_t : t Bin_prot.Size.sizer

val bin_write_t : t Bin_prot.Write.writer

val bin_writer_t : t Bin_prot.Type_class.writer

val bin_shape_t : Bin_prot.Shape.t

val t_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> t

val sexp_of_t : t -> Ppx_sexp_conv_lib.Sexp.t

val compare : t -> t -> int

include Regular.Std.Printable.S with type t := t

type t

type of printable

val to_string : t -> string

to_string x returns a human-readable representation of x

val str : unit -> t -> string

str () t is formatted output function that matches "%a" conversion format specifier in functions, that prints to string, e.g., sprintf, failwithf, errorf and, surprisingly all Lwt printing function, including Lwt_io.printf and logging (or any other function with type ('a,unit,string,...) formatN`. Example:

Or_error.errorf "type %a is not valid for %a"
  Type.str ty Exp.str exp

val pps : unit -> t -> string

synonym for str

val ppo : Core_kernel.Out_channel.t -> t -> unit

will print to a standard output_channel, useful for using in printf, fprintf, etc.

val pp_seq : Stdlib.Format.formatter -> t Core_kernel.Sequence.t -> unit

prints a sequence of values of type t

this will include pp function from Core that has type t printer, and can be used in Format.printf family of functions

include Core_kernel.Pretty_printer.S with type t := t

type t

val pp : Base.Formatter.t -> t -> unit

include Core_kernel.Comparable.S_binable with type t := t

include Core_kernel__.Comparable_intf.S_common

include Base.Comparable.S

include Base__.Comparable_intf.Polymorphic_compare

include Base.Comparisons.Infix

type t

val (>=) : t -> t -> bool

val (<=) : t -> t -> bool

val (=) : t -> t -> bool

val (>) : t -> t -> bool

val (<) : t -> t -> bool

val (<>) : t -> t -> bool

val equal : t -> t -> bool

val compare : t -> t -> int

compare t1 t2 returns 0 if t1 is equal to t2, a negative integer if t1 is less than t2, and a positive integer if t1 is greater than t2.

val min : t -> t -> t

val max : t -> t -> t

val ascending : t -> t -> int

ascending is identical to compare. descending x y = ascending y x. These are intended to be mnemonic when used like List.sort ~compare:ascending and List.sort ~cmp:descending, since they cause the list to be sorted in ascending or descending order, respectively.

val descending : t -> t -> int

val between : t -> low:t -> high:t -> bool

between t ~low ~high means low <= t <= high

val clamp_exn : t -> min:t -> max:t -> t

clamp_exn t ~min ~max returns t', the closest value to t such that between t' ~low:min ~high:max is true.

Raises if not (min <= max).

val clamp : t -> min:t -> max:t -> t Base.Or_error.t

include Base.Comparator.S with type t := t

type t

type comparator_witness

val comparator : (t, comparator_witness) Base.Comparator.comparator

include Base__.Comparable_intf.Validate with type t := t

type t

val validate_lbound : min:t Base.Maybe_bound.t -> t Base.Validate.check

val validate_ubound : max:t Base.Maybe_bound.t -> t Base.Validate.check

val validate_bound : min:t Base.Maybe_bound.t -> max:t Base.Maybe_bound.t -> t Base.Validate.check

module Replace_polymorphic_compare : Core_kernel__.Comparable_intf.Polymorphic_compare with type t := t

include Core_kernel__.Comparable_intf.Map_and_set_binable with type t := t with type comparator_witness := comparator_witness

type t

include Core_kernel.Comparator.S with type t := t

type t

type comparator_witness

val comparator : (t, comparator_witness) Core_kernel.Comparator.comparator

module Map : Core_kernel.Map.S_binable with type Key.t = t with type Key.comparator_witness = comparator_witness

module Set : Core_kernel.Set.S_binable with type Elt.t = t with type Elt.comparator_witness = comparator_witness

include Core_kernel.Hashable.S_binable with type t := t

type t

val hash_fold_t : Base.Hash.state -> t -> Base.Hash.state

val hash : t -> Base.Hash.hash_value

val hashable : t Core_kernel.Hashtbl.Hashable.t

module Table : Core_kernel.Hashtbl.S_binable with type key = t

module Hash_set : Core_kernel.Hash_set.S_binable with type elt = t

module Hash_queue : Core_kernel.Hash_queue.S with type Key.t = t

include Regular.Std.Data.S with type t := t

type t

type constructor

type info = string * [ `Ver of string ] * string option

name,Ver v,desc information attached to a particular reader or writer.

val version : string

Data representation version. After any change in data representation the version should be increased.

Serializers that are derived from a data representation must have the same version as a version of the data structure, from which it is derived. This kind of serializers can only read and write data of the same version.

Other serializers can actually read and write data independent on its representation version. A serializer, that can't store data of current version simply shouldn't be added to a set of serializers.

It is assumed, that if a reader and a writer has the same name and version, then whatever was written by the writer should be readable by the reader. The round-trip equality is not required, thus it is acceptable if some information is lost.

It is also possible, that a reader and a writer that has the same name are compatible. In that case it is recommended to use semantic versioning.

val size_in_bytes : ?⁠ver:string -> ?⁠fmt:string -> t -> int

size_in_bytes ?ver ?fmt datum returns the amount of bytes that is needed to represent datum in the given format and version

val of_bytes : ?⁠ver:string -> ?⁠fmt:string -> Regular.Std.bytes -> t

of_bytes ?ver ?fmt bytes deserializes a value from bytes.

val to_bytes : ?⁠ver:string -> ?⁠fmt:string -> t -> Regular.Std.bytes

to_bytes ?ver ?fmt datum serializes a datum to a sequence of bytes.

val blit_to_bytes : ?⁠ver:string -> ?⁠fmt:string -> Regular.Std.bytes -> t -> int -> unit

blit_to_bytes ?ver ?fmt buffer datum offset copies a serialized representation of datum into a buffer, starting from the offset.

val of_bigstring : ?⁠ver:string -> ?⁠fmt:string -> Core_kernel.bigstring -> t

of_bigstring ?ver ?fmt buf deserializes a datum from bigstring

val to_bigstring : ?⁠ver:string -> ?⁠fmt:string -> t -> Core_kernel.bigstring

of_bigstring ?ver ?fmt datum serializes a datum to a sequence of bytes represented as bigstring

val blit_to_bigstring : ?⁠ver:string -> ?⁠fmt:string -> Core_kernel.bigstring -> t -> int -> unit

blit_to_bigstring ?ver ?fmt buffer datum offset copies a serialized representation of datum into a buffer, starting from offset.

module Io : sig ... end

Input/Output functions for the given datum.

module Cache : sig ... end

Data cache.

val add_reader : ?⁠desc:string -> ver:string -> string -> t Regular.Std.reader -> unit

add_reader ?desc ~ver name reader registers a new reader with a provided name, version ver and optional description desc

val add_writer : ?⁠desc:string -> ver:string -> string -> t Regular.Std.writer -> unit

add_writer ?desc ~ver name writer registers a new writer with a provided name, version ver and optional description desc

val available_readers : unit -> info list

available_reader () lists available readers for the data type

val default_reader : unit -> info

default_reader returns information about default reader

val set_default_reader : ?⁠ver:string -> string -> unit

set_default_reader ?ver name sets new default reader. If version is not specified then the latest available version is used. Raises an exception if a reader with a given name doesn't exist.

val with_reader : ?⁠ver:string -> string -> (unit -> 'a) -> 'a

with_reader ?ver name operation temporary sets a default reader to a reader with a specified name and version. The default reader is restored after operation is finished.

val available_writers : unit -> info list

available_writer () lists available writers for the data type

val default_writer : unit -> info

default_writer returns information about the default writer

val set_default_writer : ?⁠ver:string -> string -> unit

set_default_writer ?ver name sets new default writer. If version is not specified then the latest available version is used. Raises an exception if a writer with a given name doesn't exist.

val with_writer : ?⁠ver:string -> string -> (unit -> 'a) -> 'a

with_writer ?ver name operation temporary sets a default writer to a writer with a specified name and version. The default writer is restored after operation is finished.

val default_printer : unit -> info option

default_writer optionally returns an information about default printer

val set_default_printer : ?⁠ver:string -> string -> unit

set_default_printer ?ver name sets new default printer. If version is not specified then the latest available version is used. Raises an exception if a printer with a given name doesn't exist.

val with_printer : ?⁠ver:string -> string -> (unit -> 'a) -> 'a

with_printer ?ver name operation temporary sets a default printer to a printer with a specified name and version. The default printer is restored after operation is finished.

Low level access to serializers

val find_reader : ?⁠ver:string -> string -> t Regular.Std.reader option

find_reader ?ver name lookups a reader with a given name. If version is not specified, then a reader with maximum version is returned.

val find_writer : ?⁠ver:string -> string -> t Regular.Std.writer option

find_writer ?ver name lookups a writer with a given name. If version is not specified, then a writer with maximum version is returned.