Module Std.Backend

A backend interface.

This interface must be implemented by a backend plugin, and registered with Image.register function in order to be accessible for loading images.

deprecated

Use new Ogre-powered loader interface

type perm =

| R

| W

| X

| Or of perm * perm

memory access permissions

val bin_perm : perm Bin_prot.Type_class.t

val bin_read_perm : perm Bin_prot.Read.reader

val __bin_read_perm__ : (int -> perm) Bin_prot.Read.reader

val bin_reader_perm : perm Bin_prot.Type_class.reader

val bin_size_perm : perm Bin_prot.Size.sizer

val bin_write_perm : perm Bin_prot.Write.writer

val bin_writer_perm : perm Bin_prot.Type_class.writer

val bin_shape_perm : Bin_prot.Shape.t

val compare_perm : perm -> perm -> int

val perm_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> perm

val sexp_of_perm : perm -> Ppx_sexp_conv_lib.Sexp.t

module Segment : sig ... end

A named contiguous part of file with permissions. Also, known as segment in ELF.

module Symbol : sig ... end

Symbol definition, that can span several non-contiguous parts of memory

module Section : sig ... end

Just a named region of memory.

module Img : sig ... end

A Img from a backend perspective.

type t = Core_kernel.Bigstring.t -> Img.t option

the actual interface to be implemented