Module Bil.Apply

Maps BIL operators to bitvectors.

val binop : binop -> word -> word -> word

binop op x y applies op to x and y.

before 2.5.0
precondition: the expression BinOp(op,Intx,Int y) shall be well-typed.

@after 2.5.0 if x and y have different widths then they are extended to the same width, which is the width of the largest operand. If an operator is signed, then it will be correctly sign-extended.

val unop : unop -> word -> word

unop op x applies the unary operation op to x.

val cast : cast -> int -> word -> word

cast t s x casts x using the cast type t to the given size s.

precondition: the expression Cast(t,s,Int x) shall be well-typed.