Module Bil.Storage

Predefined storage classes

deprecated
Use the Primus Framework

class linear : storage

linear storage literally implements operational semantics, but has O(N) lookup and uses space very ineffectively, as it is implemented as a list of assignments.

class sparse : storage

sparse storage is slightly more efficient storage, in comparison with linear. It uses balanced tree data structure, and provides logarithmic lookup and update method.