Module Bitvector.Trie

Prefix trees for bitvectors.

Bitvector comes with 4 predefined prefix trees:

Trie.Big.Bits - big endian prefix tree, where each token is a bit, and bitvector is tokenized from msb to lsb.

Trie.Big.Byte - big endian prefix tree, where each token is a byte, and bitvector is tokenized from most significant byte to less significant

Trie.Little.Bits - is a little endian bit tree.

Trie.Little.Byte - is a little endian byte tree.

module Big : sig ... end

module Little : sig ... end