Class Context.t

method lookup : var -> Bil.result option

self#lookup var evaluate variable var to a value that was previously bound to it. Returns None if it is unbound.

method update : var -> Bil.result -> 's

self#update var x bind variable var to value x. Returns a a context updated with the new binding.

method bindings : (var * Bil.result) seq

self#bindings returns a current list of bindings. Useful, for debugging and introspection.