Module Event.Log

Logging event.

type level =

| Debug

| Info

| Warning

| Error

type info = {

level : level;

section : string;

message : string;

}

type event += Message of info

val message : level -> section:string -> ('a, Stdlib.Format.formatter, unit) Core_kernel.format -> 'a

message level ~section fmt ... send a message of the specified level and section.

Do not use this function directly, instead include the instantiation of a Self functor, and use corresponding logging functions, e.g.,

            include Self()
            (* ... *)
            info "created some %s" "thing"

type event += Progress of {

task : string;
hierarchical task name

note : string option;
a short note

stage : int option;
entered stage

total : int option;
total number of stages

}

val progress : ?⁠note:string -> ?⁠stage:int -> ?⁠total:int -> string -> unit

progress ?note ?stage ?total name sends a progress report. This function should be used by the main components only, while plugins should use the report_progress function from the Self() interface. All parameters defaults to None.