Class Exp.finder

A visitor with a shortcut. Finder is a specialization of a visitor, that uses return as its folding argument. At any time you can stop the traversing by calling return function of the provided argument (which is by itself is a record with one field - a function accepting argument of type 'a option).

inherit 'a option Core_kernel.return visitor

method find : t -> 'a option