Class Expi.context

Context for expression evaluation.

Context provides a unique identifier for each freshly created value.

inherit Context.t

method create_undefined : 's * Bil.result

creates a fresh new result, containing an undefined value, and returns it with a modified context.

method create_word : word -> 's * Bil.result

creates a fresh new result, containing a given word, and returns it with a modified context.

method create_storage : Bil.storage -> 's * Bil.result

creates a fresh new result, containing a given storage, and returns it with a modified context.