Module type Image.Loader

An interface that a backend shall implement.

The functions provided by a loader return an OGRE document, wrapped into option and error monads, thus the three outcomes are possible with the following interpretation:

Ok None - a loader doesn't know how handle files of this type.

Ok (Some doc) - a loader was able to obtain some information from the input.

Error err - a file was corrupted, according to the loader.

val from_file : string -> Ogre.doc option Core_kernel.Or_error.t

from_file name loads a file with the given name.

val from_data : Core_kernel.Bigstring.t -> Ogre.doc option Core_kernel.Or_error.t

from_data data loads image from the specified array of bytes.