Module Interval_tree.Make_binable

Make_binable(Interval) create an abstract interval tree data type that uses abstract Interval and can be serialized via the Binable interface.

Parameters

module Interval : Interval_binable

Signature

type 'a t

include Core_kernel.Bin_prot.Binable.S1 with type 'a t := 'a t

val bin_shape_t : Bin_prot.Shape.t -> Bin_prot.Shape.t

val bin_size_t : ('a, 'a t) Bin_prot.Size.sizer1

val bin_write_t : ('a, 'a t) Bin_prot.Write.writer1

val bin_read_t : ('a, 'a t) Bin_prot.Read.reader1

val __bin_read_t__ : ('a, int -> 'a t) Bin_prot.Read.reader1

val bin_writer_t : ('a, 'a t) Bin_prot.Type_class.S1.writer

val bin_reader_t : ('a, 'a t) Bin_prot.Type_class.S1.reader

val bin_t : ('a, 'a t) Bin_prot.Type_class.S1.t

val compare : ('a -> 'a -> int) -> 'a t -> 'a t -> int

include Ppx_sexp_conv_lib.Sexpable.S1 with type 'a t := 'a t

val t_of_sexp : (Sexplib0__.Sexp.t -> 'a) -> Sexplib0__.Sexp.t -> 'a t

val sexp_of_t : ('a -> Sexplib0__.Sexp.t) -> 'a t -> Sexplib0__.Sexp.t

include S with type 'a t := 'a t with type key := Interval.t with type point := Interval.point

val sexp_of_t : ('a -> Ppx_sexp_conv_lib.Sexp.t) -> 'a t -> Ppx_sexp_conv_lib.Sexp.t

val empty : 'a t

empty x an empty interval tree

val singleton : Interval.t -> 'a -> 'a t

singleton k x creates an interval tree that has only one mapping - from the key k to data x

val least : 'a t -> Interval.point option

least t returns the least bound of the tree t.

Returns None if t is empty.

val greatest : 'a t -> Interval.point option

greatest t returns the greatest bound of the tree t.

Returns None if t is empty.

val min_binding : 'a t -> (Interval.t * 'a) option

min_bining t returns the least binding in the tree

val max_binding : 'a t -> (Interval.t * 'a) option

max_binding t returns the greatest binding in the tree

val add : 'a t -> Interval.t -> 'a -> 'a t

add t k x adds a new binding (k,x) to the mapping.

val dominators : 'a t -> Interval.t -> (Interval.t * 'a) Core_kernel.Sequence.t

dominators t k returns all intervals and their associated values that include k.

val intersections : 'a t -> Interval.t -> (Interval.t * 'a) Core_kernel.Sequence.t

intersections t k returns all intervals and their associated values that intersects with k

val intersects : 'a t -> Interval.t -> bool

intersects t k is true iff t contains an interval that intersects with k

val dominates : 'a t -> Interval.t -> bool

dominates t k is true iff all intervals in t are included in k.

val contains : 'a t -> Interval.point -> bool

contains t p is true if p belongs to at least one interval in t

val lookup : 'a t -> Interval.point -> (Interval.t * 'a) Core_kernel.Sequence.t

lookup t p returns bindings of all intervals that contain the given point

val map : 'a t -> f:('a -> 'b) -> 'b t

map k ~f maps all data values with the function f

val mapi : 'a t -> f:(Interval.t -> 'a -> 'b) -> 'b t

mapi k ~f maps all bindings with the function f

val filter : 'a t -> f:('a -> bool) -> 'a t

filter t ~f returns a tree where all elements for which f returned false are removed.

val filter_map : 'a t -> f:('a -> 'b option) -> 'b t

filter t ~f returns a tree where all elements for which f returned None are removed and all others are mapped.

val filter_mapi : 'a t -> f:(Interval.t -> 'a -> 'b option) -> 'b t

filter t ~f returns a tree where all elements for which f returned None are removed and all others are mapped.

val remove : 'a t -> Interval.t -> 'a t

remove t k removes all bindings to the key k

val remove_intersections : 'a t -> Interval.t -> 'a t

remove_intersections t k removes all bindings that intersect with the key k.

val remove_dominators : 'a t -> Interval.t -> 'a t

remove_dominators t k removes all bindings that are included (dominated by) in the interval k

val to_sequence : 'a t -> (Interval.t * 'a) Core_kernel.Sequence.t

to_sequence t returns all bindings in t

Interval Trees implement common container interface

include Core_kernel.Container.S1 with type 'a t := 'a t

val mem : 'a t -> 'a -> equal:('a -> 'a -> bool) -> bool

val length : 'a t -> int

val is_empty : 'a t -> bool

val iter : 'a t -> f:('a -> unit) -> unit

val fold : 'a t -> init:'accum -> f:('accum -> 'a -> 'accum) -> 'accum

val fold_result : 'a t -> init:'accum -> f:('accum -> 'a -> ('accum, 'e) Base__.Result.t) -> ('accum, 'e) Base__.Result.t

val fold_until : 'a t -> init:'accum -> f:('accum -> 'a -> ('accum, 'final) Base__.Container_intf.Continue_or_stop.t) -> finish:('accum -> 'final) -> 'final

val exists : 'a t -> f:('a -> bool) -> bool

val for_all : 'a t -> f:('a -> bool) -> bool

val count : 'a t -> f:('a -> bool) -> int

val sum : (module Base__.Container_intf.Summable with type t = 'sum) -> 'a t -> f:('a -> 'sum) -> 'sum

val find : 'a t -> f:('a -> bool) -> 'a option

val find_map : 'a t -> f:('a -> 'b option) -> 'b option

val to_list : 'a t -> 'a list

val to_array : 'a t -> 'a array

val min_elt : 'a t -> compare:('a -> 'a -> int) -> 'a option

val max_elt : 'a t -> compare:('a -> 'a -> int) -> 'a option