Module Std.Memmap

Memory maps. Memory map is an assosiative data structure that maps memory regions to values. Unlike in the Table, memory regions in the Memmap can intersect in an arbitrary ways. This data structure is also known as an Interval Tree.

Memmap is an instance of the Interval_tree with the Memory serving as an interval.

type 'a t = 'a memmap

memory map, aka interval trees

val sexp_of_t : ( 'a -> Ppx_sexp_conv_lib.Sexp.t ) -> 'a t -> Ppx_sexp_conv_lib.Sexp.t

val empty : 'a t

empty map

val singleton : mem -> 'a -> 'a t

singleton a memory map containing only one memory region

val min_addr : 'a t -> addr option

min_addr map is a minimum addr mapped in map

val max_addr : 'a t -> addr option

max_addr map is a maximum addr mapped in map

val min_binding : 'a t -> (mem * 'a) option

min_binding map is a minimum binding mapped in map

val max_binding : 'a t -> (mem * 'a) option

max_binding map is a maximum binding mapped in map

val add : 'a t -> mem -> 'a -> 'a t

add map mem tag adds a new memory region mem tagged with tag. If the same region was already in the map it will be tagged with the tag again, even if it has had the same tag.

val dominators : 'a t -> mem -> (mem * 'a) seq

dominators map mem an ordered sequence of all memory regions, containing mem. A memory region (x,y) contains region (p,q), iff p >= x && q <= y, where memory regions are depicted using closed intervals.

val intersections : 'a t -> mem -> (mem * 'a) seq

intersections map mem an ordered sequence of all memory regions, that intersects with mem. Memory region (x,y) intersects with region (p,q) iff there exists such z that

z >= p || z <= q && z >= x && z <= y.

In other words if there exists such byte that belongs to both memory regions.

val intersects : 'a t -> mem -> bool

intersects map mem is true if intersections map mem is not empty

val dominates : 'a t -> mem -> bool

dominates map mem if there is a non empty set of dominators

val contains : 'a t -> addr -> bool

contains map addr true if there exists such memory region mem, that Memory.contains mem addr

val lookup : 'a t -> addr -> (mem * 'a) seq

lookup map addr returns an ordered sequence of all memory containing the addr

val map : 'a t -> f:( 'a -> 'b ) -> 'b t

map m f returns a new map with each tag mapped with function f

val mapi : 'a t -> f:( mem -> 'a -> 'b ) -> 'b t

mapi m f the same as map, but f is called with two arguments: mem and tag, where mem is a memory region, and tag is a tag associated with that region.

val filter : 'a t -> f:( 'a -> bool ) -> 'a t

filter map f returns a map that contains only those elements for which f evaluated to true

val filter_map : 'a t -> f:( 'a -> 'b option ) -> 'b t

filter_map m f creates a new map by applying a function f to each tag. If f returns Some x then this region will be mapped to x in a new map, otherwise it will be dropped.

val filter_mapi : 'a t -> f:( mem -> 'a -> 'b option ) -> 'b t

filter_mapi is like filter_map but use function also accepts would associated memory region

val remove : 'a t -> mem -> 'a t

remove map mem removes all bindings to mem

val remove_intersections : 'a t -> mem -> 'a t

remove_intersections map mem removes all bindings that that intersects with mem

val remove_dominators : 'a t -> mem -> 'a t

remove_dominators map mem removes all bindings that are dominators to mem

val to_sequence : 'a t -> (mem * 'a) seq

to_sequence map converts the memmap 'a t to a sequence of key-value pairs

include Core_kernel.Container.S1 with type 'a t := 'a t

val mem : 'a t -> 'a -> equal:( 'a -> 'a -> bool ) -> bool

val length : 'a t -> int

val is_empty : 'a t -> bool

val iter : 'a t -> f:( 'a -> unit ) -> unit

val fold : 'a t -> init:'accum -> f:( 'accum -> 'a -> 'accum ) -> 'accum

val fold_result : 'a t -> init:'accum -> f:( 'accum -> 'a -> ( 'accum, 'e ) Base__.Result.t ) -> ( 'accum, 'e ) Base__.Result.t

val fold_until : 'a t -> init:'accum -> f: ( 'accum -> 'a -> ( 'accum, 'final ) Base__.Container_intf.Continue_or_stop.t ) -> finish:( 'accum -> 'final ) -> 'final

val exists : 'a t -> f:( 'a -> bool ) -> bool

val for_all : 'a t -> f:( 'a -> bool ) -> bool

val count : 'a t -> f:( 'a -> bool ) -> int

val sum : (module Base__.Container_intf.Summable with type t = 'sum) -> 'a t -> f:( 'a -> 'sum ) -> 'sum

val find : 'a t -> f:( 'a -> bool ) -> 'a option

val find_map : 'a t -> f:( 'a -> 'b option ) -> 'b option

val to_list : 'a t -> 'a list

val to_array : 'a t -> 'a array

val min_elt : 'a t -> compare:( 'a -> 'a -> int ) -> 'a option

val max_elt : 'a t -> compare:( 'a -> 'a -> int ) -> 'a option

val pp : ( Stdlib.Format.formatter -> 'a -> unit ) -> Stdlib.Format.formatter -> 'a t -> unit

pp pp_elem constracts a printer for a memmap to the given element.