Module Project.Info

Information obtained during project reconstruction.

These pieces of information are guaranteed to be discovered during the project reconstruction. See Project.create function for more information on the reconstruction process.

val file : string Bap_future.Std.stream

occurs every time a new file is opened. The value is a filename

val arch : arch Bap_future.Std.stream

occurs once input architecture is known

val data : value memmap Bap_future.Std.stream

occurs once input memory is loaded

val code : value memmap Bap_future.Std.stream

occurs once code segment is discovered

val cfg : cfg Bap_future.Std.stream

occurs every time a whole program control flow graph is changed

val symtab : symtab Bap_future.Std.stream

occurs every time a symbol table is changed

val program : program term Bap_future.Std.stream

occurs every time a program term is changed during the project reconstruction process.

val spec : Ogre.Doc.t Bap_future.Std.stream

occurs once image spec is known