Module Factory.Make

Parameters

module T : Core_kernel.T

Signature

type t = T.t

val list : unit -> string list

list source is a list of names of source providers

val find : string -> t source option

create name args finds a source provider with the given name and creates it

val register : string -> t source -> unit

register name cons registers a method that creates a given source of information. If a method with the given name already exists, then it will be superseded by a new one.