Module Std.Vector

Resizable Array.

Resizable arrays with a logarithmic push_back in the style of C++. A user need to provide a default value (c.f., DefaultConstructible requirement in C++ version).

type 'a t = 'a vector

a type of vector holding elements of type 'a

include Bin_prot.Binable.S1 with type 'a t := 'a t

type 'a t

val bin_shape_t : Bin_prot.Shape.t -> Bin_prot.Shape.t

val bin_size_t : ('a, 'a t) Bin_prot.Size.sizer1

val bin_write_t : ('a, 'a t) Bin_prot.Write.writer1

val bin_read_t : ('a, 'a t) Bin_prot.Read.reader1

val __bin_read_t__ : ('a, int -> 'a t) Bin_prot.Read.reader1

val bin_writer_t : ('a, 'a t) Bin_prot.Type_class.S1.writer

val bin_reader_t : ('a, 'a t) Bin_prot.Type_class.S1.reader

val bin_t : ('a, 'a t) Bin_prot.Type_class.S1.t

val compare : ('a -> 'a -> int) -> 'a t -> 'a t -> int

include Ppx_sexp_conv_lib.Sexpable.S1 with type 'a t := 'a t

type 'a t

val t_of_sexp : (Sexplib0.Sexp.t -> 'a) -> Sexplib0.Sexp.t -> 'a t

val sexp_of_t : ('a -> Sexplib0.Sexp.t) -> 'a t -> Sexplib0.Sexp.t

val create : ?⁠capacity:int -> 'a -> 'a t

create ?capacity default creates an empty vector with a a given capacity. It is guaranteed that the default value will never be seen by the user unless he put it into the vector explicitly with append or set.

val append : 'a t -> 'a -> unit

append xs x appends x to the end of xs

val nth : 'a t -> int -> 'a option

nth vec n returns n'th element of vector vec

val get : 'a t -> int -> 'a

get vec n like nth but raises exception if index is out of bounds

val set : 'a t -> int -> 'a -> unit

set vec n x sets n'th element of a vector vec to x if n < length vec then raises exception

val map_to_array : 'a t -> f:('a -> 'b) -> 'b array

map_to_array xs ~f copies data from xs to an array applying f to each element. See also to_array function from Container.S1 interface

val findi : 'a t -> f:(int -> 'a -> bool) -> (int * 'a) option

findi xs ~f returns an index i and a value x of the first element of xs, for which f i x is true.

val iteri : 'a t -> f:(int -> 'a -> unit) -> unit

iter xs ~f applies f i x for each x_i in xs

val foldi : 'a t -> init:'b -> f:(int -> 'b -> 'a -> 'b) -> 'b

foldi xs ~init:s_0 ~f computes f n s_n x_n, where s_n = f (n-1) s_[n-1] x_[n-1] and n is the number of elements in xs

val index : ?⁠equal:('a -> 'a -> bool) -> 'a t -> 'a -> int option

index ?equal xs x returns an index of the first element p of xs for which equal p x is true. The equal parameter defaults to the OCaml builtin polymorphic equality.

val index_exn : ?⁠equal:('a -> 'a -> bool) -> 'a t -> 'a -> int

index_exn ?equal xs x is the same as index ?equal xs x but an exception is thrown instead of None

val index_with : ?⁠equal:('a -> 'a -> bool) -> default:int -> 'a t -> 'a -> int

index_with ?equal ~default xs x same as index but returns the default value instead of None.

implements common accessors for the array, like find, fold, iter, etc

include Core_kernel.Container.S1 with type 'a t := 'a t

type 'a t

val mem : 'a t -> 'a -> equal:('a -> 'a -> bool) -> bool

Checks whether the provided element is there, using equal.

val length : 'a t -> int

val is_empty : 'a t -> bool

val iter : 'a t -> f:('a -> unit) -> unit

val fold : 'a t -> init:'accum -> f:('accum -> 'a -> 'accum) -> 'accum

fold t ~init ~f returns f (... f (f (f init e1) e2) e3 ...) en, where e1..en are the elements of t

val fold_result : 'a t -> init:'accum -> f:('accum -> 'a -> ('accum, 'e) Base.Result.t) -> ('accum, 'e) Base.Result.t

fold_result t ~init ~f is a short-circuiting version of fold that runs in the Result monad. If f returns an Error _, that value is returned without any additional invocations of f.

val fold_until : 'a t -> init:'accum -> f:('accum -> 'a -> ('accum, 'final) Base__.Container_intf.Continue_or_stop.t) -> finish:('accum -> 'final) -> 'final

fold_until t ~init ~f ~finish is a short-circuiting version of fold. If f returns Stop _ the computation ceases and results in that value. If f returns Continue _, the fold will proceed. If f never returns Stop _, the final result is computed by finish.

Example:

type maybe_negative =
  | Found_negative of int
  | All_nonnegative of { sum : int }

(** [first_neg_or_sum list] returns the first negative number in [list], if any,
    otherwise returns the sum of the list. *)
let first_neg_or_sum =
  List.fold_until ~init:0
    ~f:(fun sum x ->
      if x < 0
      then Stop (Found_negative x)
      else Continue (sum + x))
    ~finish:(fun sum -> All_nonnegative { sum })
;;

let x = first_neg_or_sum [1; 2; 3; 4; 5]
val x : maybe_negative = All_nonnegative {sum = 15}

let y = first_neg_or_sum [1; 2; -3; 4; 5]
val y : maybe_negative = Found_negative -3

val exists : 'a t -> f:('a -> bool) -> bool

Returns true if and only if there exists an element for which the provided function evaluates to true. This is a short-circuiting operation.

val for_all : 'a t -> f:('a -> bool) -> bool

Returns true if and only if the provided function evaluates to true for all elements. This is a short-circuiting operation.

val count : 'a t -> f:('a -> bool) -> int

Returns the number of elements for which the provided function evaluates to true.

val sum : (module Base__.Container_intf.Summable with type t = 'sum) -> 'a t -> f:('a -> 'sum) -> 'sum

Returns the sum of f i for all i in the container.

val find : 'a t -> f:('a -> bool) -> 'a option

Returns as an option the first element for which f evaluates to true.

val find_map : 'a t -> f:('a -> 'b option) -> 'b option

Returns the first evaluation of f that returns Some, and returns None if there is no such element.

val to_list : 'a t -> 'a list

val to_array : 'a t -> 'a array

val min_elt : 'a t -> compare:('a -> 'a -> int) -> 'a option

Returns a minimum (resp maximum) element from the collection using the provided compare function, or None if the collection is empty. In case of a tie, the first element encountered while traversing the collection is returned. The implementation uses fold so it has the same complexity as fold.

val max_elt : 'a t -> compare:('a -> 'a -> int) -> 'a option

val pp : 'a Regular.Std.printer -> 'a t Regular.Std.printer

pp pp_elem creates a vector printer that uses pp_elem to print elements.